refactor: (#328) 글 작설/수정 폼에서 카테고리 셀렉터 분리
